Add DELE RSET and NOOP commands

+++ src/main/kotlin/fr/bnancy/mail/servers/pop3/data/Pop3Session.kt
@@ -3,5 +3,5 @@ package fr.bnancy.mail.servers.pop3.data
 data class Pop3Session(
         var user: String = "",
         var currentState: Pop3SessionState = Pop3SessionState.AUTHORIZATION,
-        var messageToDelete: MutableList<Int> = mutableListOf()
+        var messageToDelete: MutableList<Long> = mutableListOf()
 )
